Agnosticity: move remix related thing to bridge
