+++ components/furnituremix/application/src/routes/$langMarket/api/cart/index.tsx
@@ -1,6 +1,6 @@
 import { ActionFunction } from '@remix-run/node';
 import { getStoreFront } from '~/core-server/storefront.server';
-import { privateJson } from '~/core-server/privateJson.server';
+import { privateJson } from '~/bridge/privateJson.server';
 import { handleAndSaveCart, hydrateCart } from '~/core-server/cart.server';
 import { getContext } from '~/core-server/http-utils.server';
